collision is working, but needs adjustment
Momentum doesn't seem to be conserved, even when I use the precision of
doubles.  The variation is on the order of +-0.0001, so it isn't really
a problem, but I'd still like to figure out why.  It may be due to
slight innaccuracies when using float velocity and masses in the
particle itself.